+++ src/utils/with-props.js
@@ -1,8 +1,8 @@
-import { cloneElement } from 'preact';
+import { toChildArray, cloneElement } from 'preact';
 
-export const withProps = (children, props) => children.map(x => {
-	if (typeof (x && x.nodeName) === 'function') {
-		return cloneElement(x, props, withProps(x.children, props));
+export const withProps = (children, props) => toChildArray(children).map(x => {
+	if (typeof (x && x.type) === 'function') {
+		return cloneElement(x, props, withProps(x.props.children, props));
 	}
 	return x;
 });